building out routes using react router 6. Learnings:  BrowserRouter, Outlet, Fragment, Link
