--- src/categories.styles.scss
@@ -1,8 +1,0 @@
-.categories-container {
-  width: 100%;
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: space-between;
-}
-
-
