+++ src/routes/shop/shop.component.jsx
@@ -1,0 +1,4 @@
+const Shop = () => (
+    <div>Shop</div>
+)
+export default Shop
